Fall_Module/requirements.txt
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
Pillow>=10.0.1
psutil  
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  
torch>=1.8.0  
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.232
pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1
mediapipe
joblib==1.3.2
scikit-learn==1.3.2
flask